packages/broker/test/unit/plugins/operator/VoteOnSuspectNodeHelper.test.ts
import {
    ParseError,
    parsePartitionFromMetadata,
} from '../../../../src/plugins/operator/VoteOnSuspectNodeHelper'

describe(parsePartitionFromMetadata, () => {
    it('throws given undefined', () => {
        expect(() => parsePartitionFromMetadata(undefined)).toThrowError(ParseError)
    })

    it('throws given invalid json', () => {
        expect(() => parsePartitionFromMetadata('invalidjson')).toThrowError(ParseError)
    })

    it('throws given valid json without field "partition"', () => {
        expect(() => parsePartitionFromMetadata('{}')).toThrowError(ParseError)
    })

    it('throws given valid json with field "partition" but not as a number', () => {
        expect(() => parsePartitionFromMetadata('{ "partition": "foo" }')).toThrowError(ParseError)
    })

    it('throws given valid json with field "partition" but outside integer range', () => {
        expect(() => parsePartitionFromMetadata('{ "partition": -50 }')).toThrowError(ParseError)
    })

    it('returns partition given valid json with field "partition" within integer range', () => {
        expect(parsePartitionFromMetadata('{ "partition": 50 }')).toEqual(50)
    })
})
